+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * fix(CommandPalette/ContextMenu/DropdownMenu): ensure items truncate work & itemTrailingIcon color
 * fix(ContentSearch): de-duplicate description and suffix
 * feat(ProseImg): improve `zoom` transition
+* feat(CommandPalette): preserve group order in search results
 
 ## [2.0.0](https://github.com/bitrix24/b24ui/compare/v1.0.4...v2.0.0) (2025-10-21)